Reimplement `StringTable` in terms of `hashbrown::HashTable`
Previously, this was using a `HashMap<String, usize>`, which requires allocating and copying individual `String`s.

Now, this uses the equivalent of a `HashSet<usize>`, where the implementation makes sure that the underlying `hash` and `eq` is using the bytes within the `buffer`. Thus no copy of individual `String`s is required, and allocations are limited to the `buffer` and `offsets`.